Bug fixes, ID system change
There was a bug where IDs could be identical. This would cause files in the cache to be overridden, this is now fixed.
(Fix: The original ID system worked by getting when the image was uploaded to the NAS, this was a timestamp in Seconds since 1970, but this system failed for the reasons stated above. It was fixed by putting together the timestamp+the size of the file (in bytes), so the chances of the IDs being identical are next to zero)